Don't exit when fopen(3) returns NULL
On failure to open a file the program should print a diagnostic
message to the standard error and move on to the next file as opposed
to exiting immediately. This allows for consistent behavior compared
to other common implementations of utilities such as cat(1) which
reduces potential confusion for the user.