Add centering by longest line (#3)
Centering by longest line requires inputting all lines before starting
output, to determine which is longest.  Use a tail queue for this to
insert at end and then iterate from start.  The lines are freed after
being output.

Unlike the static read buffer, we can't reuse the list buffers in an
easy way if we are iterating over multiple input files.  So they need to
be freed to avoid memory leaks.